+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # OSGL Tool Change Log
 
 1.18.0
+* `IO.read` stage - support ByteBuffer and byte[] #164
+* Add built-in TypeConverter between ByteBuffer and byte array #163
 * `IO` write API completeness #162
 * Provide `S.urlSafeRandom()` methods #161
 * Bean copy/mapping - it does not process `AdaptiveMap` typed source correctly #160